+++ applications/example/server/api/auth/anonymousSession.json/get.js
@@ -1,4 +1,4 @@
-﻿module.exports = function(client, callback) {
+module.exports = function(client, callback) {
   client.startSession();
   callback({ sid: client.sid });
 };