+++ applications/example/server/api/auth/register.json/post.js
@@ -1,5 +1,8 @@
-﻿module.exports = function(client, callback) {
-  application.security.register(client, client.fields.email, client.fields.password, function(err, user) {
-    callback({ result: user ? 'ok' : 'error' });
-  });
+module.exports = function(client, callback) {
+  application.security.register(
+    client, client.fields.email, client.fields.password,
+    function(err, user) {
+      callback({ result: user ? 'ok' : 'error' });
+    }
+  );
 };