+++ applications/example/server/examples/simple/endHandler.json/get.js
@@ -1,4 +1,4 @@
-﻿module.exports = function(client, callback) {
+module.exports = function(client, callback) {
   console.log('Verb handler: get.js');
   callback({ handler: 'get' });
 };