+++ applications/example/server/examples/simple/errorHandler.json/get.js
@@ -1,4 +1,4 @@
-﻿module.exports = function(client, callback) {
+module.exports = function(client, callback) {
   console.log('Verb handler: get.js (with errer to generate exception)');
   callback({ handler: 'get' });
   undefinedFunctionCall();