+++ deploy/centos6x64-src.sh
@@ -3,15 +3,15 @@ yum -y update
 yum -y install wget mc
 yum -y groupinstall "Development Tools"
 cd /usr/src
-wget http://nodejs.org/dist/v5.4.1/node-v5.4.1.tar.gz
-tar zxf node-v5.4.1.tar.gz
-rm -f ./node-v5.4.1.tar.gz
-cd node-v5.4.1
+wget http://nodejs.org/dist/v5.5.0/node-v5.5.0.tar.gz
+tar zxf node-v5.5.0.tar.gz
+rm -f ./node-v5.5.0.tar.gz
+cd node-v5.5.0
 ./configure
 make
 make install
 cd ~
-rm -rf /usr/src/node-v5.4.1
+rm -rf /usr/src/node-v5.5.0
 ln -s /usr/local/bin/node /bin
 ln -s /usr/local/bin/npm /bin
 cat >/etc/yum.repos.d/mongodb.repo <<EOL